requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==2.4.0
python-decouple==3.8
python-dotenv==1.1.1
psycopg2-binary==2.9.9
drf-spectacular==0.27.0
